src/main/java/com/pairing4good/petclinic/visit/VisitController.java
package com.pairing4good.petclinic.visit;

import com.pairing4good.petclinic.pet.Pet;
import com.pairing4good.petclinic.pet.PetRepository;
import org.springframework.stereotype.Controller;
import org.springframework.validation.BindingResult;
import org.springframework.web.bind.WebDataBinder;
import org.springframework.web.bind.annotation.*;

import javax.validation.Valid;
import java.util.Map;

@Controller
public class VisitController {

    private final VisitRepository visitRepository;
    private final PetRepository petRepository;


    public VisitController(VisitRepository visitRepository, PetRepository petRepository) {
        this.visitRepository = visitRepository;
        this.petRepository = petRepository;
    }

    @InitBinder
    public void setAllowedFields(WebDataBinder dataBinder) {
        dataBinder.setDisallowedFields("id");
    }

    @ModelAttribute("visit")
    public Visit loadPetWithVisit(@PathVariable("petId") int petId, Map<String, Object> model) {
        Pet pet = petRepository.findById(petId).get();
        model.put("pet", pet);
        Visit visit = new Visit();
        pet.addVisit(visit);
        return visit;
    }

    @GetMapping("/owners/*/pets/{petId}/visits/new")
    public String setupSave(@PathVariable("petId") int petId, Map<String, Object> model) {
        return "visits/createOrUpdateVisitForm";
    }

    @PostMapping("/owners/{ownerId}/pets/{petId}/visits/new")
    public String save(@Valid Visit visit, BindingResult result) {
        if (result.hasErrors()) {
            return "visits/createOrUpdateVisitForm";
        } else {
            visitRepository.save(visit);
            return "redirect:/owners/{ownerId}";
        }
    }

}
